% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_popepi_project.R
\name{create_popepi_project}
\alias{create_popepi_project}
\title{Utility function to set up a standardised PSB Pop Epi R project
with appropriate sub-directories and .gitignore and README.md file template
(which should be edited for each project repository)}
\usage{
create_popepi_project(path, folder_structure = TRUE, create_readme = TRUE)
}
\arguments{
\item{path}{character string specifying the path to the directory where the
new project will live.}

\item{folder_structure}{a logical value (TRUE or FALSE). If TRUE will create
sub-directories to hold input, output, and docs.}

\item{create_readme}{a logical value (TRUE or FALSE). If TRUE will create a
README.md template file with the project name as a h1 header in the content.}
}
\value{
Creates a new directory at the path specified initialised, with Rproj
setup and sub-directories, a .gitignore file and README.md file (if TRUE).
}
\description{
This function harnesses the usethis::create_project() function and adds extra
set-up for creating Git infrastructure,an R-flavoured .gitignore file, and
(optionally) appropriate sub-directories and a README.md template file.
}
\note{
As part of setting up Git infrastructure, the function will prompt the
user whether they want to make an initial commit. For some reason, the
\code{usethis} function that handles this part gives three options,
two consisting of varieties of "no" and one of "yes," and the exact text
options \strong{including their order} are different every time. Please be aware!

The "initial commit" also appears to miss some updates to the .Rproj file,
so these need to be committed by the user as part of a subsequent commit (or
you can just choose not to make an initial commit during setup and make your
first commit manually). The script template is not included in the default
initial commit, as presumably the user will want to edit and/or rename this.
}
\examples{
\dontrun{
path <- "C:/Users/username/Documents/my_new_project"
create_popepi_project(path = path, folder_structure = TRUE,
create_readme = TRUE)
}
}
