% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Inverse of \%in\% function}
\usage{
x \%nin\% table
}
\arguments{
\item{x}{vector or NULL: the values to be matched}

\item{table}{vector or NULL: the values to be matched against}
}
\value{
A \code{TRUE}/\code{FALSE} vector of the same length as \code{x}.
}
\description{
Searches whether each of element of its first argument is NOT present
in its second.
}
\examples{
x <- c("a", "b", "c")
x \%in\% letters
}
