% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_var.R
\name{my_var}
\alias{my_var}
\title{Calculates variance}
\usage{
my_var(x)
}
\arguments{
\item{x}{vector of numeric values}
}
\value{
A numeric value of the variance.
}
\description{
Calculates variance
}
\examples{
x <- c(1,2,2,4,5)
my_var(x)
}
