% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_data_cols.R
\name{validate_data_cols}
\alias{validate_data_cols}
\title{Validates that a data frame contains columns that match expected
columns.}
\usage{
validate_data_cols(data, expected_col_names)
}
\arguments{
\item{data}{data frame object: the data set to be validated}

\item{expected_col_names}{vector: character vector of expected column names to be validated against.}
}
\value{
\code{TRUE}/\code{FALSE} after completing all validations. Messages are displayed to screen if called in an interactive session.
}
\description{
This function takes a data frame and a vector of expected column names as
arguments and tests whether all the columns in the df match those listed in
the vector - both in name and number. The function returns TRUE or FALSE
invisibly (for use in downstream control statements) and if in interactive
session will print messages to the console.
}
\examples{
df <- data.frame(a = 1:3, b = 1:3, totals = c(2,4,6))
expected_col_names <- c("a", "b", "totals")
validate_data_cols(df, expected_col_names)
if(validate_data_cols(df, expected_col_names)) {
  "Dataset cols have been validated."
  }
}
